azure-functions/transcribe-audio/requirements.txt
azure-functions
azure-cognitiveservices-speech
supabase
requests
